deps: Upgrade react-intl to 5.20.1, the latest.
Like in 803852839, we restrict the version range to that single
version (instead of using a caret or a tilde), since `react-intl`
doesn't invite users to expect them to handle breaking changes
well [1].

There's no particular change we're interested in right now; we just
want to upgrade for the sake of upgrading. Nothing stands out in the
changelog [2] as being a breaking change, but we don't really trust
the package, as mentioned, so the easiest way to discover any will
probably be to have it out in a beta for a little while.

I did test switching to a different language, and things didn't seem
to break when I did that.

[1] https://github.com/formatjs/formatjs/pull/2067#issuecomment-714171574.
[2] https://github.com/formatjs/formatjs/blob/main/packages/react-intl/CHANGELOG.md