deps: Switch to expo-web-browser, from stale react-native-safari-view.
And make a libdef with flowgen.

Despite saying that the project expects to maintain compatibility
with (only) the latest version of React Native,
react-native-safari-view's README has a badge saying react-native
0.40, which is outdated by multiple years, and there's been very
little activity and few recent commits.

So, move to the Expo ecosystem, hoping it's better maintained. While
expo-web-browser is meant to work for both iOS and Android, take a
conservative approach here and only replace the uses of SafariView.

I tested social auth, a link within a message, and the "Forgot
password?" link on my iOS device.