MainTabsScreen: Turn off lazy loading.
This prevents the delay when loading a tab for the first time after app
open, which can often be significant.

This will likely add some time to the first load (since it'll render the
Streams, PMs, and other tabs for the first load), but it will cut down
on the latency when the user presses those. I wasn't able to profile the
time-to-interactive (we don't have an easy way to measure it that I've
found), but my inclination is that it's better to add a bit of time to
our (already long) TTI than it is to have switching tabs take a
significant amount of time.

What we're really prefer is to have the main screen not block on loading
the background tabs, but eagerly start loading them as soon as the app
loads. However, that's not an option that react-navigation provides.

See discussion on CZO:

https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/easy.20performance.20wins/near/1207470