StreamIcon [nfc]: Revert "Don't show as muted when unsubscribed."
This reverts commit 532ff81e8.

Before that commit, it was already the case that we don't show the
"muted" icon when not subscribed to a stream:

 * In `TitleStream` (the title in the app bar on a stream narrow),
   this is arranged by `getStreamInNarrow`.

 * In `StreamCard` (used by `StreamSettingsScreen`, this is arranged
   by the following line in the `<StreamIcon … />` call:

       isMuted={subscription ? !subscription.in_home_view : false}

 * The remaining callers don't show streams that we aren't subscribed to.

Of those, the `StreamCard` logic hadn't been working until recently;
in 9dd501406 we introduced the regrettable NULL_SUBSCRIPTION in its
caller, and that caused the `subscription ? …` to take the truthy
branch, treating the bogus NULL_SUBSCRIPTION as real data.  But we
fixed that in d963d36cc by removing that use of NULL_SUBSCRIPTION.

So this didn't have any functional effect.

This component's interface is certainly kind of a mess.  Instead of
taking `isMuted` and `isPrivate` separately, and instead of requiring
each caller to have the right logic for handling unsubscribed
streams, it'd be better for this component to take whole Stream and
Subscription objects -- or possibly just a stream ID, and do its own
`connect` to get the data from Redux -- and handle that logic
for itself.

But adding one more prop that's also ultimately derived from
subscription data doesn't really make things any cleaner; at least it
doesn't while the callers still have their own copies of the
unsubscribed-vs-muted logic.