unread [nfc]: Add `getUnread` selector.
This will let us start passing around the unread state as a whole,
instead of its various pieces like the part for streams or 1:1 PMs.

That in turn will help us start to better encapsulate the specifics
of the unread state's data structures, which will make it easier to
modify them.