logging: Use Sentry's captureMessage and captureException in logToSentry.
captureException takes an Error object, and captureMessage takes a
string. It makes sense to use captureException when we have an
Error, then, and captureMessage when we have a string -- instead of
coercing both into a string and synthesizing a stack trace to
attach.

Greg said, about the comment "the static API's `captureException`
doesn't allow passing strings",

> [...] I think the idea of that comment is that we want to pass an
> `Error` when we have an `Error`, and just a string when we have
> just a string.
>
> And then implicitly it's looking for one method to use in both
> cases. Neither of those works, so they were both rejected.
>
> But the conditional seems like a fine solution to that problem.