generateInboundEvents: Remove early return on sending a 'content' event.
Greg finds that the early return was added in b5fb47765, as (it
would seem) a small optimization rather than a significant fix or
feature.

Fundamentally, more than one interesting change can happen at the
same time, between `prevProps` and `nextProps`. We probably can't
rule out React (or a future React) chunking what we might've
expected to be several separate updates into a single update, and we
don't want to miss anything important by returning early before we
can process all the changes we need to. So, make the control flow
more symmetrical among the different types of updates.

Also (and this is relevant for #3851), the 'content' event
duplicates some of the work of the 'fetching' event: it sometimes
takes charge of removing the message placeholders, and that seems
like a job we shouldn't have to rely on it doing. So, allow the
'fetching' event to get sent, instead of returning early before it
can. Now we'll be free to change or replace the 'content' event with
something that doesn't hide the message placeholders.

See discussion at
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/message-list.20diffing/near/1218348.

Related: #3851