android notif [nfc]: Remove `InstanceIDListenerService.java`.
This is deprecated [1][2] in favor of overriding
`FirebaseMessagingService#onNewToken`.

While not causing any noticable bugs, its good to have up to date
code. One benefit of this specific commit is that it removes some
code.

Something related but not done in this commit is to update
firebase-messaging itself, which as of now stands at 22.0.0 .

[1] https://firebase.google.com/docs/reference/android/com/google/firebase/iid/FirebaseInstanceIdService
[2] https://firebase.google.com/support/release-notes/android#firebase_instance_id_version_1620