android notif [nfc]: Remove `ByConversationMap`.
`ByConversationMap` was originally intended to store conversations
based on `Identity` as key, so that we would be able to group
conversations by realmId easily.

Here we are removing it, in favor of using
`NotificationManager#getActiveNotifications` in the following
commits.

Notifications will persist automatically if we are to do this, and
we don't have to be concerned with handling logic for storing
past notifications, Android SDK will abstract it away for us.

This will eventually be followed by removal of`ConversationMap`,
and other associated code.

Skips: #4633