createPersistor: Stop blending snapshots when we initiate writes.
This has the downside that, if a write is in progress, state updates
will be ignored until there's a subsequent update after the write is
complete; that subsequent update will begin a new write. This isn't
ideal (and we don't want to keep it this way forever), but it's
better than not persisting anything at all, or persisting data that
doesn't correspond to a value that an entire in-memory state object
had at some point.

We'll fix that in the upcoming commits, but in the meantime, this
makes the code much easier to nudge into a reasonably performant
algorithm that uses `storage.multiSet` instead of several
`storage.setItem`s.