ComposeBox: Switch message editing to use onSend callback.
This may have a regression that `blur` is not called on the message
input box when a message is done being edited, but as far as I can tell,
a bug was preventing that blur call from doing anything in the first
place.

This will also fix a bug that we send a "typing started" notification
but never a "typing stopped" notification when editing a message.