api types [nfc]: Say $Exact<…> when spreading ApiResponseSuccess.
We've had this as an exact object type, but the real type is
inexact.  When spreading it, though, we always just want to
incorporate the properties it specifically has.  The way to
spell that is `$Exact<ApiResponseSuccess>`.

This has no effect at the moment because the type is exact in the
first place, but it'll let us change it to inexact and keep getting
the intended results.