api errors [nfc]: Add comment on error-text dependency.
I just happened across this line in our Sentry setup code, for
reasons unrelated to the API-errors refactoring I did the other day.
I didn't change this text then, but I might have, and probably would
have missed that that other spot needed to be updated to match.
Add a comment to help if someone does change the text later.