persist [nfc]: Remove conditionals for serializer / deserializer config
These don't need to be defined conditionally. We pass these in our
config, and we don't expect to stop.

`json-stringify-safe`, an NPM package with its last publish six
years ago, is removed, as its only caller was in one of the idle
branches of the conditional. After removing this, and reordering the
'invariant' import line, we can now remove the `import/order` eslint
override, too.