sharing types [nfc]: Tweak comments on SharedData.
The admonition to keep these in sync with platform-native code
is addressed to someone editing this definition, rather than
someone looking at its interface in order to use it elsewhere,
so it naturally goes in an implementation comment (with `//` or
plain `/* … */`) rather than a jsdoc comment (with `/** … */`).

Putting it more immediately next to the very implementation also
improves the chances that someone editing the latter will actually
notice the comment and act on it.

Also add a quick note about iOS to go with the note about the
situation on Android.