mute [nfc]: Clarify a confusing shouldBeMuted case, in tests and code.
The test says the message is muted if the stream is *not* in the
"mute list"?  Is that backward? … Ah, no, the point is that this
covers when the stream isn't in the *subscriptions* list -- i.e.,
you're not subscribed.

Clarify that in the code, with a comment on why this case isn't just
an error case that shouldn't happen, and why it gets this behavior;
and fix the test description too.

Also add a test for the basic happy case, where you *are* subscribed
and it isn't muted.  That's what the next few cases are implicitly
comparing against -- it wouldn't be very interesting that they
returned false if this one didn't return true.