mute [nfc]: Move part of shouldBeMuted's logic up into its caller.
There are lots of different types of narrows, and for many of them
we really want to do nothing at all here.  So this logic will be
made somewhat easier to reason through by pushing those conditions
up to the top level of the getShownMessagesForNarrow control flow.

As a first step in that, move these two conditions, which are easy to
move because they don't interact with the specific messages at all.