SmartUrlInput: Use `useFocusEffect`.
Following the doc at
  https://reactnavigation.org/docs/use-focus-effect/.

Greg points out [1] that `useFocusEffect` seems to have the effect
of calling the callback even on the component's first mount (if the
screen is focused), whereas the old listener-based code would not.
That doesn't seem to have a practical effect, here; empirically, it
looks like the input gets focused anyway before this change. That's
probably due to the `autoFocus` prop we pass to the `TextInput`.

[1] https://github.com/zulip/zulip-mobile/pull/4914#discussion_r675919671.