msglist tests: Make messageUpdates tests well-typed.
This in fact fixes one test case, which was passing while not
actually exercising the intended logic: in the first test case
for getMessageUpdateStrategy, `prevProps.narrow` had been missing
even though `nextProps.narrow` was present, causing `sameNarrow`
to be false.  That means we go through the "switching narrows"
control-flow path, and don't reach the `allNewMessages` condition
which this case is intended to exercise.