notif: Remove call to cancelAll everywhere.
Whenever a user checks out a notification it will dismiss every
notification. Same thing happens with call to `onClear`.

This is not the usual expected behaviour and infact it would be
much better to cancle only the notification that we tap on.

This commit makes the relevant changes for that.