MentionWarnings [nfc]: Convert to function component.
This one's a little unusual because of the design where imperative
methods are exposed for a parent of MentionWarnings to call.

Use `forwardRef` and `useImperativeHandle`, as the Hooks way to
implement this design; see
  https://reactjs.org/docs/hooks-reference.html#useimperativehandle.

Note that "forwardRef" already appears near the bottom of
MentionWarnings.js. There, we're telling `connect` to forward the
ref to the component wrapped by the HOC. That one will disappear in
the next commit, when we stop using `connect`. But the second one
will remain, since that's our ticket to allowing callers to take a
`ref` to our function component; see the doc linked above, and
  https://reactjs.org/docs/refs-and-the-dom.html#refs-and-function-components.