flow: Properly pull in @react-native-async-storage/async-storage libdef.
We'd been trying to do this by putting a line in `[libs]`, under the
following comment:

```
; If the following files are not explicitly named, Flow ignores them
; in favor of their autogenerated versions in flow-typed/npm.
```

But the problem isn't that it's getting shadowed by an autogenerated
stub in flow-typed/npm. The problem is that the Flow types that the
library ships with aren't very good, and we want to ignore them and
use our own libdef. For an example, see `getItem` in
node_modules/@react-native-async-storage/async-storage/lib/commonjs/AsyncStorage.js.flow:

```
  static getItem(key: string, callback?: Function): Promise<*> {
    return createPromise(() => {
      return window.localStorage.getItem(key);
    }, callback);
  }
```

So, do a thing that accomplishes that.