redux-persist: Improve logging of errors.
`logging.warn` has the best chance of helpfully reporting an Error
object if we pass it as the first argument. In some of these cases,
we weren't even passing an Error at all, when we had one available
-- so, do so, and have it be the first argument.

The second argument should be a plain, JSONable object. So, make
that happen, too, so we don't get Sentry events with things like

  "extra": {
    "0": "u",
    "1": "s",
    "2": "e",
    "3": "r",
    "4": "G",
    "5": "r",
    "6": "o",
    "7": "u",
    "8": "p",
    "9": "s"
  }