user [nfc]: Comment on possible future getHaveServerData checks.
I don't see an immediate way we could check right now whether any
server data we have is actually for the current active account, or
for some other account (presumably for a previous active account,
and left over due to our consistency-lacking data storage model.)

The reasoning in the comment below this, about why it's mostly fine
for this function not to spot inconsistencies between pieces of the
server data, probably also holds between the server data and accounts.
But there are also some plausible future changes, including at least
one we actually intend to make soon anyway, that would make it easy
to add some such checks, and at that point it'd be nice to do so.
So add a comment about those.