register [nfc]: Rename three plain Redux actions
In particular:

  INITIAL_FETCH_START -> REGISTER_START
  INITIAL_FETCH_ABORT -> REGISTER_ABORT
  REALM_INIT          -> REGISTER_COMPLETE

We now have this set of three actions that clearly report what's
going on with the `/register` call.

`/register` gives us an event queue ID, and soon we'd like to
improve our Redux state logic for storing that. Those changes should
now be easier to reason through.

Also, make a few minor NFC tweaks so it's clearer that these three
are related, like changing the order of some imports.