ZulipBanner [nfc]: Better support single-row layout (1/3).
The spec [1] has a special case that we haven't been handling at
all: when the text is one line and there's just one button, the text
and the button should go on the same row, instead of getting split
up into two rows as we've been doing.

When the text is very short, it does indeed look like we're wasting
a lot of space by putting the text and the buttons on separate rows.
We should do something to avoid that feeling of wasted space, in the
spirit of that case. We haven't found a perfect way of doing this;
the solution in this sub-series of commits will emphasize code
simplicity and improvement over the status quo.

First, with continued respect for the case of two-line text with two
buttons, in the spec (and this is NFC): refactor how we set the
required distance of 36 between the baseline of the text's first
line and the top of the container. Where before we had
`lineHeight: 20` plus `marginTop: 16`, instead give 10px of that
`marginTop` to the container's `paddingTop`.

That will make it easier to follow the one-line, one-button case in
the spec: the illustration there says there should be 10px between
the top of the action button and the top of the container.

[1] https://material.io/components/banners#specs