usePrevious [nfc]: Default to undefined instead of null.
The main motivation for this is that it makes the types behave more
uniformly between the case where `initValue` is or isn't provided:
this way, the value we actually use for it is always the value passed,
rather than sometimes being that and sometimes of a different type.

That in turn will let us tighten the types so that when we have a
real initial value to provide, the return type becomes simply `T`
rather than `T | null` or `T | void`.