android: Report logged warnings and errors to Sentry.
The one for malformed FCM messages may be especially informative
as we start treating FCM messages from very old servers (that lack
important pieces of information) as malformed.

More generally this covers all of our call sites to `Log.w` and
above.  We should use `Zlog.w` and `Zlog.e` in place of those for
any future call sites, too -- the `Log` call gets it to the device
log, and that could be quite handy for development, but for issues
encountered in a published build the Sentry call will get it to
our Sentry dashboard, which is much more likely to be helpful.

This change drops the message strings that were at each of these
specific call sites.

 * In Sentry, there isn't a good place to put those -- the exception
   type and message are the most salient pieces of information -- and
   because that's our most valuable channel for debugging, it didn't
   seem like a good idea to have the `ZLog` API accept a message that
   we weren't going to be able to make good use of in Sentry.  That's
   why we drop them.

 * In the device log, passing a throwable to `Log.*` causes the
   throwable's class name and message to get printed, as well as the
   stack trace.  In each of these cases, that should cover all the
   information that was in these messages anyway.  In particular,
   the most informative of these messages are really just restating
   the name of the class that we already know the exception is an
   instance of.  So we're not losing anything material by dropping
   them, even when reading the device log rather than Sentry.