android notif [nfc]: Update comment on `user_id` in payload.
The server release that introduced this, 2.1, is nearly two years old
and it'd be reasonable for us to ignore FCM messages that lack it.

But we just recently (in 37074ce79, released in v27.169, which went
out to users in general on 2021-09-09) started showing users a warning
banner if they're using a 2.0.x server, so it'd be kind to give a bit
of time for people to act on that if their server admins have
procrastinated upgrading.  And, conveniently, we don't immediately
have a reason we *need* to require this field, so there's no loss in
leaving it be for now.

Anyway, for now just update the comment with exact pointers to the
developments that it forecast but were in the future when it was
written.