flags reducer: Handle `all: true` events correctly.
We'd been implicitly assuming that the only case of these is
to mark all as read.  That's the only case that server versions
to date actually send; but there's no guarantee of that in the API,
and we should handle them correctly in general in case other such
events are sent in the future -- for example, perhaps "unstar all".

See CZO discussion:
https://chat.zulip.org/#narrow/stream/378-api-design/topic/UPDATE_MESSAGE_FLAGS.20.60all.3A.20true.60/near/1222451