import: Revert '.' -> './' workaround.
This reverts commit 24f1a17a1.

The underlying issue in Metro is supposed to have been fixed in
Metro 0.60: https://github.com/facebook/metro/pull/540.

`yarn why metro` shows that we're using Metro v0.64, now that we've
upgraded RN.

Inspecting the bundle after this change confirms that the imports
are working as expected.

See discussion at
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/Notification.20Crash/near/1215541.