webview html [nfc]: Simplify show-date-separator logic a bit
This is a bit simpler at the top by just having fewer boolean
operations and one fewer reference to the truthiness of prevMessage.

It's a bit neater at showSender, because the reason to show the
sender isn't really that the date is different, in itself -- it's
that we just showed a date separator, and so we need to re-affirm
the sender because that's visually a tighter grouping.

To see this is NFC, consider first the case where prevMessage
is falsy, and then where it's truthy.