webview html: Tie "show sender" to "show recipient header"
This is more accurately what we mean at showSender: we show the
sender because we showed a recipient header, just like we do
because we showed a date separator.  If the actual recipient
changed but we didn't show a recipient header, there'd be no
point in showing the sender again, not for that reason anyway.

This would be NFC as long as we always do show a recipient header
when the recipient changes, which would be true as long as the
recipient never changes when canShowRecipientHeaders is false.

And that is *almost* true: we only make canShowRecipientHeaders
false in conversation narrows, and the *actual* recipient never
changes within one of those.  But there is a small bug in
isSameRecipient (see comment there for details) where an Outbox
message and a proper Message can have isSameRecipient false even
though they're actually to the same PM thread.

So this change actually fixes one symptom of that bug: if you're
looking at a PM thread with an Outbox message that has such a
mismatch, then previously we would show the sender anew, and now
we don't.  (The bug still has another symptom: if you go visit
an interleaved narrow where we *do* show recipient headers,
we'll show one for that mismatch, and duly show the sender anew
to go with it.)