redux types: Mark most selectors as per-account vs. global
After this change, the bulk of our selectors are correctly marked
as per-account, taking a PerAccountState -- which is most of
them -- or global, taking a GlobalState.

The major class of exceptions is the `state.accounts` selectors
found in `accountsSelectors`, as discussed in a previous commit.

A more boring class of exceptions is the many small anonymous
selectors (mostly as callbacks to `useSelector`) that get bits of
the settings and session data.  Many of those are really
per-account, but for now we leave them all marked as global.

At this stage none of our thunk actions or React components get
converted.  That means they're all still acting on GlobalState; which
is why it's very convenient that (for the next while) we've set up
PerAccountState so that it's a supertype of GlobalState, and a value
of the latter can be freely implicitly used as the former.

(Of all the Redux-related names we'll be splitting, this one is
unusual in requiring renames through the per-account majority of the
app code.  Most will be like Selector in that the existing name is
short and generic and we'll keep that for the per-account majority
to use.  GlobalState is the main exception, because the existing
name is explicitly "global".)

This doesn't yet have any real effect, because PerAccountState is
for the moment just an alias of GlobalState.  The real test that we've
assigned each of these to the proper column will be that Flow passes
when we make those two types distinct, coming soon.