redux types: Use GlobalSelector on global selectors
We're in the process of marking selectors as either per-account,
taking a PerAccountState, or global, taking a GlobalState.

In this commit, we apply the GlobalSelector alias where it's needed.
In a couple of commits, we'll address selectors that don't use that
alias; they currently all say GlobalState, and we'll switch to
PerAccountState where appropriate.

We actually go farther here than we ultimately want to: we mark all
the `state.accounts` selectors found in `accountsSelectors` as
global.  In fact these are a mix of global and per-account, mostly
per-account; but teasing those apart will take some careful
sorting-out of their callers' expectations, so we'll leave that to a
bit later in this series of PRs.  For now we leave them all marked
as global, which is the conservative option in that it means Flow
notices anytime we call one of them from a per-account context.

This doesn't yet have any real effect, because PerAccountState is
for the moment just an alias of GlobalState.