presence [nfc]: Drop a tryGetAuth for plain getAuth, like other actions
This thunk action reportPresence is in much the same situation as
most of our other thunk actions: it's meant to act on a single
account's data; it acts on the active account's data; it has a
bug where if the active account changes while its network request
is pending, it can apply the response to the wrong account's data;
and it can't do much if there is no logged-in active account.

It had been unusual in that it checked at the top for a logged-in
active account, and bailed out silently if there was none.  That's
because it's unusual in that it's called from a spot that starts
operating before the user logs in, and keeps operating if they
log out.

But ever since we overhauled the presence-reporting code in
6efc4497a, that caller (PresenceHeartbeat) already checks whether
there's a logged-in active account, and only invokes this action if
there is.  So it can go ahead and make that assumption exactly the
same way that the bulk of our thunk actions do.  That will simplify
bringing it along with the rest when we make them act on a specific
account they're dispatched for.

Also add a couple of related notes for that future migration.