account [nfc]: Mark getIdentity as per-account
With an assumeSecretlyGlobalState, because for the moment getAuth
is still marked as global.  (That one has a lot of callers, so
we'll get to it later.)

Also add a TODO that perhaps this function shouldn't care if the
account is logged in.  But that's the existing behavior and it
seems orthogonal to what we're doing now, so we leave it be.