account [nfc]: Strength-reduce various getActiveAccount calls
All of these are in per-account contexts: the main UI, or a
server fetch (indeed the initial fetch.)

None of them need the whole Account object.  That object includes
the API key, which is a good reason not to pass it around where
we can avoid it; and it has both the user's identity on the server
and miscellaneous bits like the server version, which might get
rearranged to separate places in the future, so it's convenient
to put those through separate indirections like getIdentity and
getServerVersion.