presence [nfc]: Make choice of newUserInput explicit; add TODO
I don't really know what this parameter means, so can't evaluate if
false is the right value for it.  The history doesn't seem to be
enlightening; we've had it this way since early on.

Also make the other argument `isActive` non-optional.  The existing
call site passes a value for it explicitly, and this will keep it
that way.