user: Split getHaveServerData as per-account vs global
This function has a bit of logic that makes sense only in a global
context (asking whether there even is an account to have state for),
but then the bulk of it is per-account.

One group of its callers is firmly per-account, in the initial fetch
in fetchActions.js.  So those would like this function to be
per-account.

But then its other caller is withHaveServerDataGate.  It's possible
that should also be per-account, but I'm not sure of that -- in
particular I'm not sure we only use it in contexts where there is an
account at all.  Or possibly that should be the case but it would take
some refactoring to make it so.  For the present, then, I'd rather
defer that question by not changing its behavior; I want to just keep
it making all the same checks it currently does.

To accommodate both sides of that, we split this into a global and
a per-account version, and have each caller use the appropriate
one.  This means those per-account actions in fetchActions.js are
now calling the per-account version of this selector.