redux types: Start using GlobalThunkAction / GlobalDispatch
This marks as global all our global thunk actions and some of
their callers.

The remaining callers are a couple of Redux components.  For now,
they still invoke these thunk actions with a plain `Dispatch`,
and so we rely for now on the fact that we haven't yet made that
distinct from `GlobalDispatch` -- the former still provides a
`GlobalState` to its actions.  In the next commit, we'll make it
possible to mark those components as global so that they get a
`GlobalDispatch`, and then we can tighten `Dispatch`.