react-redux types: Propagate `dispatch` type properly, with BoundedDiff
In the old version with `$Diff`, Flow wouldn't complain if a
component expected a `dispatch` prop with too specific a type for
what it was actually going to get: that is, if the type it expected
was a subtype of the type actually provided by `connect`, encoding
more-specific expectations about the value than `connect` was
actually promising to meet.

It actually *would* complain if the component didn't have specific
*enough* expectations of the type.  That is, it'd accept calling
`connect` on a component expecting a `dispatch: empty` -- something
`connect` couldn't possibly give it -- but reject doing so with
`dispatch: mixed`, which should be perfectly fine.  Effectively,
by using `$Diff` here, we seem to have gotten the comparison backward.
(And Flow is unable to check our work in doing so, because the whole
reason we have this file is that the react-redux libdef is too lax.)

That hasn't really been a practical issue to date, because the only
reasonable type for something called `dispatch` in our codebase has
been `Dispatch`.  But we're soon going to have two flavors of that
type, one for acting on a particular account's data and one for
acting on all accounts' or purely account-independent data, so the
distinction will become an important one.

Happily, our `BoundedDiff` solves this problem.