deps: Upgrade flow-coverage-report to 0.8.0, the latest.
No breaking changes are announced:
  https://github.com/rpl/flow-coverage-report/releases

Running `yarn test:flow-coverage` before and after this change,
the HTML reports it generates are visually identical in the
browser.  The only detectable difference is that the new version
runs about twice as fast.

For more on using this tool, see:
  https://github.com/zulip/zulip-mobile/pull/4955#issuecomment-934614690
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Dependabot/near/1203387.

Apart from it being nice in general to be on the latest versions
of things, this also removes all five low-severity vulnerabilities
reported by `yarn audit` that implicate flow-coverage-report.

[greg: tested, revised commit message]