redux types: Introduce PerAccountState, for now just an alias
Also add a small function assumeSecretlyGlobalState, to explicitly
coerce a PerAccountState to GlobalState; we'll do that in a few
places as scaffolding while we're converting things over.

And add a tiny type-test that converting in the other direction can
be done implicitly.  In the middle of the transition, we'll be doing
that a *lot* -- for example at most places we call a selector from a
thunk action or a React component, when we've converted selectors but
not yet thunk actions or React components -- so it'll be quite handy
that it can be one-way implicit like that.