notif: Explicitly get active account state in narrowToNotification
This is a global thunk action -- it starts by explicitly looking at
all the accounts -- but then invokes some per-account selectors,
meaning to do so on the active account's per-account state.  Make
that transition explicit, so that we can stop letting a GlobalState
masquerade as a PerAccountState.

Marking that transition also brings up that we should explicitly
deal with the case where there is no active account (because there
is no account.)