presence [nfc]: Add fixme on hasAuth gate in PresenceHeartbeat
This really is a per-account component.  In particular it really
wants a per-account `dispatch`, for the part where it actually
reports presence to the server.  That means it's correctly using
the per-account form of `connect`, which will soon start giving
it only a PerAccountState.

There are several possible ways we could reconcile that with this
line where it's using the global selector `getHasAuth`; one is in
the comment added below.  But for now, just mark this so we know
we'll need to resolve it before we actually physically distinguish
the global state from a per-account state.