webview [nfc]: Give message-list elements a new '.msglist-element' class
This marker indicates that a given element is a proper "message-list
element", i.e., the DOM counterpart of the MessageListElement
objects we pass around on the React Native side.

We give this marker to those "message-list elements" but not to
other elements.

See getMessageNode in js.js for a function that returns one of these
elements in the ancestry of a given element. We'll soon refactor
that function so that it uses `.closest('.msglist-element')` to get
the same element.