narrow [nfc]: Pass stream IDs to Narrow constructors, almost everywhere
This is all made possible by the preceding series of changes where
we parsed, tracked, or otherwise made stream IDs available in a
variety of places where we'd previously had only a stream name.

This leaves just a handful of call sites not passing IDs -- in
particular, where the stream we want to refer to comes itself from
some Narrow value.  Those we'll take care of together with actually
storing the stream ID in the Narrow.