generics: Fix variance on IsSupertype
Flow shouldn't have accepted the variance annotation we had here;
it did so only because of the bug discovered in the previous commit.

Fortunately, the discussion there gives us a framework for seeing
what variance annotations would actually be legitimate.  Use those.

And happily it turns out we weren't relying anywhere on the old,
wrong, annotation.