generics: Use BoundedDiff with read-only arguments, as specced
Recently in 50d1aa6e8 we edited this type's jsdoc to say the type
arguments are expected to be not only exact but read-only object
types, because it seemed complicated to accurately describe how it
behaves when they're not.

But outside the type's own tests, we've been observing that constraint
in exactly zero of its call sites, oops.  Fix those.