sharing: Fix a race on sending vs. changing the destination
After we make the send the request and while we're awaiting its
completion, the user could further edit the inputs, in particular
the stream and topic.  In that case `this.props.sendTo` will change,
and when we get to `onShareSuccess` we'd end up using the new
`sendTo`, and so navigating to the conversation identified by the
new values, which isn't the one we actually sent the message to.

Instead, pass to `onShareSuccess` the same `sendTo` value we
actually just used for sending.

Separately, it'd probably be good to have the UI prevent editing
the inputs while the request is pending, because there's no good
way for that to behave.  Even with that, though, it'd be best to
pass the data explicitly like this so that it's clear from local
examination that the same `sendTo` is used throughout, rather than
have to verify that we're 100% effective at preventing mutation.