gitattributes: Suppress diffs on Jest snapshot files
Or at least the ones we have now, which can have many-hundred-line
diffs when we make a one-line change to the HTML for a message.

Quite possibly this is what we'd want for all Jest snapshot files,
but if we find we want it on future files we can expand it then.