android notif: Log explicitly if getToken fails
This occasionally fails, and if it does we get an event for it in
Sentry because the exception propagates.  Luckily the exception
doesn't interfere with anything else, because after this point
there's nothing else in the `start` function, and its caller
doesn't await its result.

But better to make that the case by construction: log an error,
and then skip just the relevant step (of trying to post the gotten
token to the Zulip server(s)) and move on without an exception.
That way if we add more at the bottom of this function, or do some
other refactor, we won't end up with the app failing to start
properly because an exception here aborted our startup work.

Also, try to discern whether it looks like it looks like the user
is hitting this exception transiently or persistently -- i.e.,
whether it's blocking them from getting notifications -- and
distinguish the log event accordingly.

Fixes: #5061