android build: Bump compileSdkVersion to 30, aka Android 11
This version has been out for a while! We could totally have bumped
this earlier. But now we have to do so, because we plan to bump
`targetSdkVersion`, and

  minSdkVersion ≤ targetSdkVersion ≤ compileSdkVersion

should hold.

This is not related to which versions we support for running the app.
See comments and the further info linked from there for details.

Also garbage-collect an outdated comment about .travis.yml. We moved
from Travis to GitHub Actions in 7dfa1c2cb, and so far our GitHub
Actions config doesn't care about this value.