api [nfc]: Use the greppable `TODO(server-N.M)` form consistently.
We've recently started to actually drop old conditionals that added
complexity and were needed for supporting old versions of the server.

To do more of that, it's good to have a systematic format we use when
marking that something depends on the server version, so that as we
desupport an old version we can come sweep through and find all the
new opportunities to simplify.

We've generally tried to make these references reasonably clear, but
hadn't had a concise form to use consistently.  A year ago in
e47aa004b, I introduced a form of comments like this:
    // TODO(server-2.0): switch to numeric user IDs, not emails.
and I think that works well.  Sweep through now, using a variety of
less-precise search patterns to try to find existing comments of this
kind, and convert them all to that standard form.

Many of these refer to server versions we already don't support --
with ServerCompatBanner we tell users their server is unsupported if
it's older than 2.1.  So a lot of these are opportunities we can
already take.  But we'll leave those for separate commits.

Also make small related comment fixes:
 * The projected "2.2" was renamed 3.0 before release.
 * Drop a TODO that isn't a TODO.