account [nfc]: Let getServerVersion return a non-null type
I added a new selector here in 172764ce8, just after making it the
case (in 31e5da3fc) that this information is always present when we
have server data.  I was focusing on the feature level, rather than
the version number, and apparently didn't notice that there was
already an existing selector for the version number.  So I had the
new selector use that new invariant, but the existing selector's
existing callers didn't get the benefit of it.

Fix that, and simplify the callsites accordingly.  All three of them
already assume we have server data.

We consolidate on the name getServerVersion, rather than
getZulipVersion, because the mobile app itself is of course also
"Zulip" and has its own version number.  The name "Zulip version"
appears in the server's own API because in that context it makes
sense, and it's leaked into parts of our codebase from there, but
we can refrain from propagating it further.