api types: Cut Message#sender_domain, removed in server 1.6
Servers that sent this property are now long obsolete -- 1.6.0 was
released 2017-06-06, well over four years ago -- so there's not likely
any debugging value in even keeping it around commented out.

This appears to be our only TODO(server-1.6).  We're now fully
upgraded to take advantage of the Zulip Server 1.6.0 release!