api [nfc]: Always pass apply_markdown, and reflect that in types
We were already always doing this: in get-messages we already
expressed it in the binding itself, and in register-queue we did so
at its only caller.

Move this choice inside the bindings in the latter case too.  That
lets us simplify the `Message` type... at which the type-checker
tells us that some example data needs to be updated.  So that
example data demonstrates that it definitely wasn't previously clear
to us that for us the `content_type` is always HTML, never Markdown,
because we'd written it there the other way.

And now that the one callsite isn't passing any options here, also
take the opportunity to make the options object optional entirely.

The change to example data causes a couple of Jest snapshot files
to need to be updated.  Do that by passing `--updateSnapshot` to Jest.