android notif: Wordsmith the group-PM heading a bit
This is somewhat shorter and I think conveys all the same information.
It also skips the term "group PM", which we do use in user-facing text
but is a little bit jargony and nice to avoid where we can make the
meaning clear directly.

Also drop the final periods "."; in the UI this string ends up getting
shown with a colon after it, so the period is redundant.